tstannotate

+++ src/ZChop.jl
@@ -4,17 +4,17 @@ export zchop
 
 const zchopeps = 1e-14
 
-function zchop{T<:Real}(x::T, eps::Float64=zchopeps)
+function zchop{T<:Real}(x::T, eps=zchopeps)
     abs(x) > eps ? x : zero(T)
 end
 
-function zchop{T<:Complex}(x::T, eps::Float64=zchopeps)
+function zchop{T<:Complex}(x::T, eps=zchopeps)
     rx = real(x)
     ix = imag(x)
     complex(zchop(rx,eps),zchop(ix,eps))
 end
 
-function zchop{T<:Number}(a::AbstractArray{T}, eps::Float64=zchopeps)
+function zchop{T<:Number}(a::AbstractArray{T}, eps=zchopeps)
     b = similar(a)
     for i in 1:length(a)
         av = a[i]
@@ -23,7 +23,7 @@ function zchop{T<:Number}(a::AbstractArray{T}, eps::Float64=zchopeps)
     b
 end
 
-zchop(x,eps::Float64=zchopeps) = map((x)->zchop(x,eps),x)
+zchop(x,eps=zchopeps) = map((x)->zchop(x,eps),x)
 
 end # module ZChop
